Fix DPCM offset issue with sound engine.

+++ FamiStudio/Source/Project/Project.cs
@@ -1511,12 +1511,16 @@ namespace FamiStudio
             }
         }
 
-        public void Cleanup()
+        public void Cleanup(bool removeUnusedSamples = true)
         {
             DeleteUnusedInstruments();
-            UnmapUnusedSamples();
-            DeleteUnmappedSamples();
             DeleteUnusedArpeggios();
+
+            if (removeUnusedSamples)
+            {
+                UnmapUnusedSamples();
+                DeleteUnmappedSamples();
+            }
         }
 
         public bool MergeProject(Project otherProject)
